requirements.txt
matplotlib
networkx
numpy
plotly
scipy
